spark-docker-example/requirements.txt
pyspark===2.4.3
kafka-python
python-keycloak
requests
